+++ src/main/java/me/kimihiqq/vouchermanagement/domain/voucher/dto/VoucherDto.java
@@ -1,0 +1,4 @@
+package me.kimihiqq.vouchermanagement.domain.voucher.dto;
+
+public record VoucherDto(String type, long discount) {
+}
