+++ src/main/java/me/kimihiqq/vouchermanagement/domain/voucher/repository/MemoryVoucherRepository.java
@@ -1,7 +1,7 @@
-package me.kimihiqq.vouchermanagement.repository;
+package me.kimihiqq.vouchermanagement.domain.voucher.repository;
 
 import lombok.extern.slf4j.Slf4j;
-import me.kimihiqq.vouchermanagement.domain.Voucher;
+import me.kimihiqq.vouchermanagement.domain.voucher.Voucher;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
 import java.util.*;
