+++ src/main/java/me/kimihiqq/vouchermanagement/domain/voucher/service/VoucherService.java
@@ -1,0 +1,11 @@
+package me.kimihiqq.vouchermanagement.domain.voucher.service;
+
+import me.kimihiqq.vouchermanagement.domain.voucher.Voucher;
+import me.kimihiqq.vouchermanagement.domain.voucher.dto.VoucherDto;
+
+import java.util.List;
+
+public interface VoucherService {
+    Voucher createVoucher(VoucherDto voucherDto);
+    List<Voucher> listVouchers();
+}
