--- src/main/java/me/kimihiqq/vouchermanagement/dto/VoucherDto.java
@@ -1,4 +1,0 @@
-package me.kimihiqq.vouchermanagement.dto;
-
-public record VoucherDto(String type, long discount) {
-}
