--- src/main/java/me/kimihiqq/vouchermanagement/service/VoucherService.java
@@ -1,11 +1,0 @@
-package me.kimihiqq.vouchermanagement.service;
-
-import me.kimihiqq.vouchermanagement.domain.Voucher;
-import me.kimihiqq.vouchermanagement.dto.VoucherDto;
-
-import java.util.List;
-
-public interface VoucherService {
-    Voucher createVoucher(VoucherDto voucherDto);
-    List<Voucher> listVouchers();
-}
